AI Agents/langgraph-agents/rag-agent/pyproject.toml
[project]
name = "rag-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.26",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.0",
    "pypdf>=5.6.1",
    "python-dotenv>=1.1.1",
]
